# Conflicts:
#	docs/deployment/production.md
#	packages/back-end/src/app/controllers/easy-genomics/laboratory/run/process-update-laboratory-run.lambda.ts
#	packages/back-end/src/infra/stacks/easy-genomics-api-stack.ts
#	packages/back-end/src/infra/stacks/easy-genomics-nested-stack.ts
#	packages/back-end/test/app/controllers/easy-genomics/laboratory/run/create-laboratory-run.lambda.test.ts
#	packages/back-end/test/app/controllers/easy-genomics/laboratory/run/process-update-laboratory-run.lambda.test.ts
#	packages/back-end/test/app/controllers/easy-genomics/laboratory/update-laboratory.lambda.test.ts
#	packages/back-end/test/infra/stacks/easy-genomics-nested-stack.test.ts
#	packages/front-end/src/app/repository/modules/labs.ts
#	packages/shared-lib/src/app/openapi/easy-genomics-api.json
#	packages/shared-lib/src/app/openapi/easy-genomics-api.yaml
#	packages/shared-lib/src/app/schema/easy-genomics/laboratory-run.ts
#	packages/shared-lib/src/app/schema/easy-genomics/laboratory.ts
#	packages/shared-lib/src/app/types/easy-genomics/generated.d.ts
#	packages/shared-lib/src/app/types/easy-genomics/laboratory-run.d.ts
#	packages/shared-lib/src/app/types/easy-genomics/laboratory.d.ts

+++ packages/back-end/src/app/controllers/easy-genomics/data-collections/remove-files-from-sample.lambda.ts
@@ -16,7 +16,7 @@ export const handler: Handler = async (
     if (!RemoveFilesFromSampleSchema.safeParse(body).success) throw new InvalidRequestError();
 
     const { userId, laboratory } = await assertSequenceCollectionsAccess(event, body.LaboratoryId);
-    sampleService.assertBucketMatchesLab(laboratory, body.S3Bucket);
+    await sampleService.assertLaboratoryHasS3BucketAccess(laboratory, body.S3Bucket);
 
     await sampleService.removeFilesFromSample(laboratory, userId, body.S3Bucket, body.SampleId, body.Keys);
     return buildResponse(200, JSON.stringify({ ok: true }), event);
